refactor: remove barrel file, use direct module imports
Remove src/platform/keybindings/index.ts as per project guidelines
that prohibit barrel files within src/. Update all consumers to
import directly from concrete modules.

Amp-Thread-ID: https://ampcode.com/threads/T-019c0c78-bead-76da-9ea5-a2874fd8f1f4
