--- src/platform/keybindings/index.ts
@@ -1,7 +1,0 @@
-export { CORE_KEYBINDINGS } from './defaults'
-export { KeyComboImpl } from './keyCombo'
-export { KeybindingImpl } from './keybinding'
-export { useKeybindingService } from './keybindingService'
-export { useKeybindingStore } from './keybindingStore'
-export type { KeyCombo, Keybinding } from './types'
-export { zKeybinding } from './types'
